tests: move WordBox/Segment builders to tests/helpers
- Add tests/helpers.py and tests/__init__.py; import from tests.helpers
- Slim conftest.py to fixtures only (tmp_job_root)
- Update test_e055, test_e251, test_raster, test_e142 to use tests.helpers

Made-with: Cursor
